refactor(AGenericService, RepositoryResponse): simplify response handling
Update response logic in AGenericRepository and AGenericService to improve clarity and maintainability. Change the not found response in AGenericRepository and refactor multiple methods in AGenericService to utilize a new ToBusinessResponse method for handling repository responses.

Bump version in BaseUtility.csproj from 2.1.1 to 2.1.2.
Add new overload for ToBusinessResponse in RepositoryResponseExtension to support IEnumerable mapping.