chore(release): update version to 7.1.0 and enhance ValidationPipeline
Update the project file `BaseUtility.csproj` to change the version
number from `7.0.0` to `7.1.0`. The target framework remains set to
`.NET 8`, with other properties unchanged.

Add a new method `AddHandler` in the `ValidationPipeline` class to
allow users to add a single `IValidationHandler` to the internal
list of handlers, enhancing flexibility in handler management.